requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.0.0
textblob==0.17.1
pymongo==4.3.3
